Overhaul of Series class (and resulting clean-up of other classes) -Series and ProbabilityDensity now inherit from History class; other structures for saving results can be added by subclassing History. These structures can be multiply inherited by the Differential class to combine their features.
-Some functions where renamed or aliased to make a consistent calling
interface for all history structures (e.g. line_of_data -> update).
More consistency: all time management (storing start, end of simulation
and related tests/functions), which are common to all child history
structures, now done in History class.

-Initial conditional is also now entirely dealt with by the Series class
(derived from History), instead of stored in Differential and copied
over, further reducing coupling.

-The differential class no longer requires a reference to the history
structure (as it did a reference to Series): it is simply passed as an
argument when evaluating drift and diffusion. This reduces coupling and
allows to remove the dubious odeSeriesRef variable from the integrator

-*STILL TO FIX*: The differential's parameters are not updated correctly
(the ParameterMap is updated, but not the differential's variables
linked by pointer)